Enable symbolic Jacobian and W factorizations
Enables ModelingToolkit to use the generated Jacobian and W-fact function for precomputed symbolic factorizations of I-gamma*J to make the implicit ODE solvers really fly. Usually this gives around a 2x speedup, so we'll see what we get here. The factorization can take awhile to compute on really large systems, so I capped it at 16 at which it no longer applies this optimization.