Fix and test callbacks (i.e. fix mtime) (#831)
* Fix and test callbacks

The continuous callbacks is what is referred to as `mtime`. This is a general form for controlling the differential equation solver and setting events externally. The issue was that the callbacks were accidentally sent to `solve` which overrode the dosing callbacks and thus broke the core event handling. By just manually merging and sending the merged callbacks, everything works as planned.  Fixes https://github.com/PumasAI/Pumas.jl/issues/804

* remove plotting from tests

* add back splat