Unify parallelism interfaces under ensemblealg (#842)
* Unify parallelism interfaces under ensemblealg

This makes the estimation interface the same as the simulation interface. In some sense it's a little peculiar since simobs is actually using the EnsembleProblem while estimation algorithms are not, but the implementation of these three algorithms actually matches, i.e. EnsembleSerial is just a loop, EnsembleThreads is just an `@threads` loop, and EnsembleDistributed is just a `pmap`, so they are actually doing the same thing in these two cases. Therefore, it's nice to just make the interface the same. This still leaves off EnsembleSplitThreads and EnsembleGPUArray which would require a lot more work, but a nice error that those are not implemented yet is thrown, so I think this is a good spot for now as a single parallelism API but with clear ways to move forward.

* fix parallelism test

* fix error handling test